issue with deleting object containing object brick
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@280 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/models/Object/Class/Data/Objectbricks.php
@@ -272,7 +272,9 @@ class Object_Class_Data_Objectbricks extends Object_Class_Data
     public function delete($object)
     {
         $container = $this->load($object);
-        $container->delete($object);
+        if($container) {
+            $container->delete($object);
+        }
     }
 
     public function getAllowedTypes()
